chore: use os specific build tags for the os specific code.
Also run go-lint against each supported OS.

This should reduce binary sizes a bit.
