+++ internal/nexodus/keys.go
@@ -22,61 +22,6 @@ const (
 	privateKeyPermissions = 0600
 )
 
-// handleKeys will look for an existing key pair, if a pair is not found this method
-// will generate a new pair and write them to location on the disk depending on the OS
-func (ax *Nexodus) handleKeys() error {
-	switch ax.os {
-	case Darwin.String():
-		publicKey := readKeyFile(ax.logger, darwinPublicKeyFile)
-		privateKey := readKeyFile(ax.logger, darwinPrivateKeyFile)
-		if publicKey != "" && privateKey != "" {
-			ax.wireguardPubKey = publicKey
-			ax.wireguardPvtKey = privateKey
-			ax.logger.Infof("Existing key pair found at [ %s ] and [ %s ]", darwinPublicKeyFile, darwinPrivateKeyFile)
-			return nil
-		}
-		ax.logger.Infof("No existing public/private key pair found, generating a new pair")
-		if err := ax.generateKeyPair(darwinPublicKeyFile, darwinPrivateKeyFile); err != nil {
-			return fmt.Errorf("Unable to locate or generate a key/pair: %w", err)
-		}
-		ax.logger.Debugf("New keys were written to [ %s ] and [ %s ]", darwinPublicKeyFile, darwinPrivateKeyFile)
-		return nil
-
-	case Windows.String():
-		publicKey := readKeyFile(ax.logger, windowsPublicKeyFile)
-		privateKey := readKeyFile(ax.logger, windowsPrivateKeyFile)
-		if publicKey != "" && privateKey != "" {
-			ax.wireguardPubKey = publicKey
-			ax.wireguardPvtKey = privateKey
-			ax.logger.Infof("Existing key pair found at [ %s ] and [ %s ]", windowsPublicKeyFile, windowsPrivateKeyFile)
-			return nil
-		}
-		ax.logger.Infof("No existing public/private key pair found, generating a new pair")
-		if err := ax.generateKeyPair(windowsPublicKeyFile, windowsPrivateKeyFile); err != nil {
-			return fmt.Errorf("Unable to locate or generate a key/pair: %w", err)
-		}
-		ax.logger.Debugf("New keys were written to [ %s ] and [ %s ]", windowsPublicKeyFile, windowsPrivateKeyFile)
-		return nil
-
-	case Linux.String():
-		publicKey := readKeyFile(ax.logger, linuxPublicKeyFile)
-		privateKey := readKeyFile(ax.logger, linuxPrivateKeyFile)
-		if publicKey != "" && privateKey != "" {
-			ax.wireguardPubKey = publicKey
-			ax.wireguardPvtKey = privateKey
-			ax.logger.Infof("Existing key pair found at [ %s ] and [ %s ]", linuxPublicKeyFile, linuxPrivateKeyFile)
-			return nil
-		}
-		ax.logger.Infof("No existing public/private key pair found, generating a new pair")
-		if err := ax.generateKeyPair(linuxPublicKeyFile, linuxPrivateKeyFile); err != nil {
-			return fmt.Errorf("Unable to locate or generate a key/pair: %w", err)
-		}
-		ax.logger.Debugf("New keys were written to [ %s ] and [ %s ]", linuxPublicKeyFile, linuxPrivateKeyFile)
-		return nil
-	}
-	return nil
-}
-
 // generateKeyPair a key pair and write them to disk
 func (ax *Nexodus) generateKeyPair(publicKeyFile, privateKeyFile string) error {
 	cmd := exec.Command(wgBinary, "genkey")
