+++ internal/nexodus/route.go
@@ -1,53 +1,5 @@
 package nexodus
 
-import (
-	"net"
-
-	"github.com/nexodus-io/nexodus/internal/models"
-)
-
-// handlePeerRoute when a new configuration is deployed, delete/add the peer allowedIPs
-func (ax *Nexodus) handlePeerRoute(wgPeerConfig wgPeerConfig) {
-	switch ax.os {
-	case Darwin.String():
-		// Darwin maps to a utunX address which needs to be discovered (currently hardcoded to utun8)
-		devName, err := getInterfaceByIP(net.ParseIP(ax.wgLocalAddress))
-		if err != nil {
-			ax.logger.Debugf("failed to find the darwin interface with the address [ %s ] %v", ax.wgLocalAddress, err)
-		}
-		// If child prefix split the two prefixes (host /32) and child prefix
-		for _, allowedIP := range wgPeerConfig.AllowedIPs {
-			_, err := RunCommand("route", "-q", "-n", "delete", "-inet", allowedIP, "-interface", devName)
-			if err != nil {
-				ax.logger.Debugf("no route deleted: %v", err)
-			}
-			if err := AddRoute(allowedIP, devName); err != nil {
-				ax.logger.Debugf("%v", err)
-			}
-		}
-
-	case Linux.String():
-		for _, allowedIP := range wgPeerConfig.AllowedIPs {
-			routeExists, err := RouteExists(allowedIP)
-			if err != nil {
-				ax.logger.Warnf("%v", err)
-			}
-			if !routeExists {
-				if err := AddRoute(allowedIP, ax.tunnelIface); err != nil {
-					ax.logger.Errorf("route add failed: %v", err)
-				}
-			}
-		}
-
-	case Windows.String():
-		for _, allowedIP := range wgPeerConfig.AllowedIPs {
-			if err := AddRoute(allowedIP, ax.tunnelIface); err != nil {
-				ax.logger.Debugf("route add failed: %v", err)
-			}
-		}
-	}
-}
-
 func (ax *Nexodus) addChildPrefixRoute(childPrefix string) {
 
 	routeExists, err := RouteExists(childPrefix)
@@ -62,42 +14,5 @@ func (ax *Nexodus) addChildPrefixRoute(childPrefix string) {
 
 	if err := AddRoute(childPrefix, ax.tunnelIface); err != nil {
 		ax.logger.Infof("error adding the child prefix route: %v", err)
-	}
-}
-
-// handlePeerRoute when a peer is this handles route deletion
-func (ax *Nexodus) handlePeerRouteDelete(dev string, wgPeerConfig models.Device) {
-	switch ax.os {
-	case Darwin.String():
-		// TODO: OSX route lookups
-		for _, allowedIP := range wgPeerConfig.AllowedIPs {
-			if err := DeleteRoute(allowedIP, dev); err != nil {
-				ax.logger.Debug(err)
-			}
-		}
-
-	case Linux.String():
-		for _, allowedIP := range wgPeerConfig.AllowedIPs {
-			routeExists, err := RouteExists(allowedIP)
-			if !routeExists {
-				continue
-			}
-			if err != nil {
-				ax.logger.Debug(err)
-			}
-			if routeExists {
-				if err := DeleteRoute(allowedIP, dev); err != nil {
-					ax.logger.Debug(err)
-				}
-			}
-		}
-
-	case Windows.String():
-		// TODO: Windoze route lookups
-		for _, allowedIP := range wgPeerConfig.AllowedIPs {
-			if err := DeleteRoute(allowedIP, dev); err != nil {
-				ax.logger.Debug(err)
-			}
-		}
 	}
 }
